Fix problem with Int on 32 bit devices, it is by default 32 bit not 64bit

+++ QBTracker/QubitSDK/APIClient/Entities/QBEventEntity.swift
@@ -47,7 +47,7 @@ struct QBEventEntity {
         self.session = session
     }
     
-    mutating func setBatchTs(batchTs: Int) {
+    mutating func setBatchTs(batchTs: Int64) {
         self.meta?.batchTs = batchTs
     }
